Change that startTime gets set when getInstanceDescriptor() in GamesHandler gets called

+++ android/tinySSB/app/src/main/java/nz/scuttlebutt/tremolavossbol/tssb/games/GameInstance.kt
@@ -15,7 +15,7 @@ class GameInstance(gameType: String, fid: String, identity: SSBid) {
     var game : Game? = null
     var participantFid : String = "-"
     var ownerFid : String = "-"
-    var startTime : Int = 0
+    var startTime : Long = 0
     var state : GameStates = GameStates.STOPPED
 
     init {
